+++ dtls/src/flight/flight5.rs
@@ -709,6 +709,7 @@ async fn initalize_cipher_suite(
             &h.algorithm,
             &h.signature,
             &state.peer_certificates,
+            cfg.insecure_verification,
         ) {
             return Err((
                 Some(Alert {