+++ webrtc/src/api/setting_engine/mod.rs
@@ -65,6 +65,7 @@ pub struct SettingEngine {
     pub(crate) sdp_media_level_fingerprints: bool,
     pub(crate) answering_dtls_role: DTLSRole,
     pub(crate) disable_certificate_fingerprint_verification: bool,
+    pub(crate) allow_insecure_verification_algorithm: bool,
     pub(crate) disable_srtp_replay_protection: bool,
     pub(crate) disable_srtcp_replay_protection: bool,
     pub(crate) vnet: Option<Arc<Net>>,
@@ -246,6 +247,11 @@ impl SettingEngine {
         self.disable_certificate_fingerprint_verification = is_disabled;
     }
 
+    /// allow_insecure_verification_algorithm allows the usage of certain signature verification
+    /// algorithm that are known to be vulnerable or deprecated.
+    pub fn allow_insecure_verification_algorithm(&mut self, is_allowed: bool) {
+        self.allow_insecure_verification_algorithm = is_allowed;
+    }
     /// set_dtls_replay_protection_window sets a replay attack protection window size of dtls_transport connection.
     pub fn set_dtls_replay_protection_window(&mut self, n: usize) {
         self.replay_protection.dtls = n;